[ADD-UNIT-TEST]: Fix gRPC integration test.
