+++ backend/src/test/java/com/example/demo/BaseGrpcIntegrationTest.java
@@ -1,15 +1,5 @@
 package com.example.demo;
 
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
-import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.context.ActiveProfiles;
+public abstract class BaseGrpcIntegrationTest extends BaseIntegrationTest {
 
-@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT, classes = DemoApplication.class)
-@DirtiesContext
-@ActiveProfiles({"integration-test"})
-public abstract class BaseGrpcIntegrationTest {
-
-    @LocalServerPort private int port;
 }
