+++ backend/src/test/java/com/example/demo/BaseRestIntegrationTest.java
@@ -1,22 +1,12 @@
 package com.example.demo;
 
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.web.client.RestTemplateBuilder;
-import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.context.annotation.Bean;
 import org.springframework.http.converter.protobuf.ProtobufJsonFormatHttpMessageConverter;
 import org.springframework.test.context.ActiveProfiles;
 
-@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT, classes = DemoApplication.class)
-@ActiveProfiles({"integration-test", "no-security"})
-public abstract class BaseRestIntegrationTest {
-
-    @LocalServerPort private int port;
-
-    public int getPort() {
-        return port;
-    }
+@ActiveProfiles({"no-security"})
+public abstract class BaseRestIntegrationTest extends BaseIntegrationTest {
 
     @Bean
     public RestTemplateBuilder restTemplateBuilder() {
